Avoid imports between tests files
