+++ packages/salesforce-adapter/e2e_test/custom_object_instances.test.ts
@@ -27,7 +27,7 @@ import { runFiltersOnFetch, createElement, removeElementAndVerify, createInstanc
 import { apiName, isInstanceOfCustomObject } from '../src/transformers/transformer'
 import customObjectsFilter from '../src/filters/custom_objects'
 import customObjectsInstancesFilter from '../src/filters/custom_objects_instances'
-import { createCustomSettingsObject } from '../test/filters/custom_settings_filter.test'
+import { createCustomSettingsObject } from '../test/utils'
 import { LIST_CUSTOM_SETTINGS_TYPE } from '../src/constants'
 
 /* eslint-disable @typescript-eslint/camelcase */
