+++ packages/workspace/test/workspace/static_files/functions.test.ts
@@ -18,7 +18,7 @@ import { StaticFile, Value } from '@salto-io/adapter-api'
 import { Functions } from '../../../src/parser/functions'
 import { getStaticFilesFunctions } from '../../../src/workspace/static_files/functions'
 import { StaticFilesSource } from '../../../src/workspace/static_files'
-import { mockStaticFilesSource } from './common.test'
+import { mockStaticFilesSource } from '../../utils'
 
 describe('Functions', () => {
   let functions: Functions
